requirements.txt
# fastapi
# uvicorn
# groq
# langchain
# langchain-community
# langchain-text-splitters
# langchain-huggingface
# langchain-groq
# faiss-cpu
# PyPDF2
# sentence-transformers
# python-dotenv
# numpy
# python-multipart



# ---------- Core ----------
python-dotenv==1.0.1
pydantic==1.10.13

# ---------- FastAPI ----------
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# ---------- LangChain Core ----------
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4



# ---------- LLM (Groq) ----------
langchain-groq==0.1.9

# ---------- Embeddings ----------
sentence-transformers==2.6.1
transformers==4.41.2
torch==2.2.2
langchain-huggingface==1.2.0


# ---------- Vector DB ----------
faiss-cpu==1.8.0

# ---------- PDF ----------
pypdf==4.2.0

# ---------- Utils ----------
numpy==1.26.4







